Remove CloudKit references in preparation for conversion to Firebase
